fix(review): bot review followups
- _text.py: tighten safe_json_loads to actually return None for empty
  dict/list and non-container values, matching the docstring contract
  on both the strict and repair passes (Gemini).
- memory/_sqlite.py: extract open_sqlite_wal into a leaf module so
  microverse.ops.metrics no longer pulls in microverse.memory and its
  WorldContext dependency just to open a SQLite db (CodeRabbit).
- memory/_sqlite.py: surface actual_mode in the WAL-failure error
  message instead of the raw SQLite row tuple (Gemini).

CodeRabbit's "EpisodicMemory needs a Lock" finding intentionally not
applied — episodic.py:81 documents the pre-existing autocommit-per-write
serialization design; this PR is simplification-only and adding locks
is a behavior change with potential perf implications.

All 212 tests still green; ruff clean; format clean.
