+++ src/microverse/memory/__init__.py
@@ -17,9 +17,7 @@ single-model loop where we control prompt shape.
 
 from __future__ import annotations
 
-import sqlite3
 import time
-from pathlib import Path
 from typing import TYPE_CHECKING
 
 from microverse.agents.base import WorldContext
@@ -30,26 +28,6 @@ if TYPE_CHECKING:
 
 
 SEVEN_DAYS_S: float = 7 * 24 * 3600.0
-
-
-def open_sqlite_wal(path: str | Path) -> sqlite3.Connection:
-    """Open a SQLite connection with the project's standard durability
-    pragmas (WAL + ``synchronous=NORMAL``).
-
-    Strict by default: file-backed paths that fail to enter WAL raise.
-    The ``:memory:`` carve-out exists because in-memory dbs always report
-    ``"memory"`` from ``PRAGMA journal_mode`` — no durability surface to
-    defend, so don't fail. ``check_same_thread=False`` so the watchdog
-    can read while the tick loop writes; callers serialize logically.
-    """
-    conn = sqlite3.connect(str(path), check_same_thread=False)
-    mode_row = conn.execute("PRAGMA journal_mode=WAL").fetchone()
-    actual_mode = str(mode_row[0]).lower() if mode_row else ""
-    if str(path) != ":memory:" and actual_mode != "wal":
-        conn.close()
-        raise RuntimeError(f"failed to enable WAL on db {path!r}; got mode={mode_row}")
-    conn.execute("PRAGMA synchronous=NORMAL")
-    return conn
 
 
 def est_tokens(text: str) -> int:
